address review: reuse traffic policy merge helpers
Move ShallowCopyTrafficPolicy and MergeTrafficPolicy into pilot/pkg/model
so the backend-policy merge reuses them instead of duplicating the field
logic. They cannot live in pilot/pkg/networking/util because that package
imports model. util now delegates to the model versions.

Also extend the test to cover Tls and RetryBudget being set by both the
DestinationRule and the backend policy, asserting the DestinationRule wins
on the conflicting fields while the backend fills the unset ones.
