+++ pilot/pkg/networking/core/cluster_builder.go
@@ -658,7 +658,7 @@ func (cb *ClusterBuilder) buildInboundCluster(clusterPort int, bind string,
 			opts.policy = &networking.TrafficPolicy{}
 		} else {
 			// copy policy to prevent mutating the original destinationRule trafficPolicy
-			opts.policy = util.ShallowCopyTrafficPolicy(opts.policy)
+			opts.policy = model.ShallowCopyTrafficPolicy(opts.policy)
 		}
 		opts.policy.ConnectionPool = sidecarConnPool
 	}
