+++ pilot/pkg/networking/util/util.go
@@ -848,38 +848,9 @@ func MergeSubsetTrafficPolicy(original, subsetPolicy *networking.TrafficPolicy, 
 
 	// merge DR with subset traffic policy
 	// Override with subset values.
-	mergedPolicy := ShallowCopyTrafficPolicy(original)
+	mergedPolicy := model.ShallowCopyTrafficPolicy(original)
 
-	return mergeTrafficPolicy(mergedPolicy, subsetPolicy, hasPortLevel)
-}
-
-// Note that port-level settings will override the destination-level settings.
-// Traffic settings specified at the destination-level will not be inherited when overridden by port-level settings,
-// i.e. default values will be applied to fields omitted in port-level traffic policies.
-func mergeTrafficPolicy(mergedPolicy, subsetPolicy *networking.TrafficPolicy, hasPortLevel bool) *networking.TrafficPolicy {
-	if subsetPolicy.ConnectionPool != nil || hasPortLevel {
-		mergedPolicy.ConnectionPool = subsetPolicy.ConnectionPool
-	}
-	if subsetPolicy.OutlierDetection != nil || hasPortLevel {
-		mergedPolicy.OutlierDetection = subsetPolicy.OutlierDetection
-	}
-	if subsetPolicy.LoadBalancer != nil || hasPortLevel {
-		mergedPolicy.LoadBalancer = subsetPolicy.LoadBalancer
-	}
-	if subsetPolicy.Tls != nil || hasPortLevel {
-		mergedPolicy.Tls = subsetPolicy.Tls
-	}
-
-	if subsetPolicy.Tunnel != nil {
-		mergedPolicy.Tunnel = subsetPolicy.Tunnel
-	}
-	if subsetPolicy.ProxyProtocol != nil {
-		mergedPolicy.ProxyProtocol = subsetPolicy.ProxyProtocol
-	}
-	if subsetPolicy.RetryBudget != nil {
-		mergedPolicy.RetryBudget = subsetPolicy.RetryBudget
-	}
-	return mergedPolicy
+	return model.MergeTrafficPolicy(mergedPolicy, subsetPolicy, hasPortLevel)
 }
 
 func shadowCopyPortTrafficPolicy(portTrafficPolicy *networking.TrafficPolicy_PortTrafficPolicy) *networking.TrafficPolicy {
@@ -891,22 +862,6 @@ func shadowCopyPortTrafficPolicy(portTrafficPolicy *networking.TrafficPolicy_Por
 	ret.LoadBalancer = portTrafficPolicy.LoadBalancer
 	ret.OutlierDetection = portTrafficPolicy.OutlierDetection
 	ret.Tls = portTrafficPolicy.Tls
-	return ret
-}
-
-// ShallowCopyTrafficPolicy shallow copy a traffic policy, portLevelSettings are ignored.
-func ShallowCopyTrafficPolicy(original *networking.TrafficPolicy) *networking.TrafficPolicy {
-	if original == nil {
-		return nil
-	}
-	ret := &networking.TrafficPolicy{}
-	ret.ConnectionPool = original.ConnectionPool
-	ret.LoadBalancer = original.LoadBalancer
-	ret.OutlierDetection = original.OutlierDetection
-	ret.Tls = original.Tls
-	ret.Tunnel = original.Tunnel
-	ret.ProxyProtocol = original.ProxyProtocol
-	ret.RetryBudget = original.RetryBudget
 	return ret
 }
